refactor(mcp): keep body-cap constant on a leaf Edge import
Move MAX_JSON_RPC_BODY_BYTES out of the MCP constants/upgrade graph so
docs-mcp and mcp-proxy only pull the number, and cover understated
Content-Length streaming overflow in the shared reader tests.
